transfer functions from "users.controller" to "users.service"
