+++ src/controller/users.controller.ts
@@ -1,24 +1,41 @@
 import express, { NextFunction, Request, Response } from "express";
-import usersService from "../service/users.service";
 import bcrypt from "bcryptjs";
-import debug from "debug";
 const router = require("express").Router();
 import Joi from "@hapi/joi";
 import UsersService from "../service/users.service";
 
-const log: debug.IDebugger = debug("app:users-controller");
 class UsersController {
   async createUser(req: Request, res: Response): Promise<Response> {
     req.body.password = await bcrypt.hash(req.body.password, 10);
-    const userId = await usersService.create(req.body);
+    const userId = await UsersService.create(req.body);
     return res.status(201).send({ id: userId });
+  }
+
+  async authenticateSchema(req: Request, res: Response, next: NextFunction) {
+    const schema = Joi.object({
+      login: Joi.string().required(),
+      password: Joi.string().required(),
+    });
+    validateRequest(req, res, next, schema);
+  }
+
+  async authenticate(
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> {
+    const { login, password } = req.body;
+    UsersService.authenticate(login, password)
+      .then((...user) => {
+        res.json(user);
+      })
+      .catch(next);
   }
 }
 
-export default new UsersController();
-
-export async function validateRequest(
+async function validateRequest(
   req: Request,
+  res: Response,
   next: NextFunction,
   schema: any
 ) {
@@ -29,36 +46,15 @@ export async function validateRequest(
   };
   const { error, value } = schema.validate(req.body, options);
   if (error) {
-    next(
-      `Validation error: ${error.details.map((x: any) => x.message).join(", ")}`
-    );
+    return res.status(401).json({
+      message: `Validation error: ${error.details
+        .map((x: any) => x.message)
+        .join(", ")}`,
+    });
   } else {
     req.body = value;
     next();
   }
 }
 
-export async function authenticateSchema(
-  req: Request,
-  res: Response,
-  next: NextFunction
-) {
-  const schema = Joi.object({
-    login: Joi.string().required(),
-    password: Joi.string().required(),
-  });
-  validateRequest(req, next, schema);
-}
-
-export async function authenticate(
-  req: Request,
-  res: Response,
-  next: NextFunction
-) {
-  const { login, password } = req.body;
-  UsersService.authenticate_function(login, password)
-    .then((...user) => {
-      res.json(user);
-    })
-    .catch(next);
-}
+export default new UsersController();
